requirements.txt
python-docx
pdfplumber
requests
